incorporate review requests

+++ src/s2/util/math/exactfloat/bignum.cc
@@ -110,7 +110,7 @@ std::optional<Bignum> Bignum::FromString(absl::string_view s) {
   bool negative = false;
 
   // Consume optional +/- at the front.
-  const char* begin = s.data();
+  const char* begin = s.data();  // std::from_chars needs pointers, not iterators
   if ((*begin == '+' || *begin == '-')) {
     negative = (s[0] == '-');
     ++begin;
@@ -118,8 +118,7 @@ std::optional<Bignum> Bignum::FromString(absl::string_view s) {
 
   const char* end = s.data() + s.size();
   while (begin < end) {
-    size_t chunk_len =
-        std::min(static_cast<size_t>(end - begin), kMaxChunkDigits);
+    size_t chunk_len = std::min(end - begin, kMaxChunkDigits);
     Bigit chunk = 0;
     auto result = std::from_chars(begin, begin + chunk_len, chunk);
     if (result.ec != std::errc() ||
